mod: Changed license type from BSL to AGPL
I wasn't familiar with AGPL licenses. My original goal was and still is to protect the significant amount of work I've put into the project from commercial competition but more than one person has complained that its 'not true open source' and a helpful one suggested the concept of going with AGPL. I read the terms and they're a good tradeoff. I want to court open source activity and if this is how I've gotta do it then so be it.